+++ _source/alias/media.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+#r
+if [[ -f /usr/bin/cyanrip  ]];
+then
+  alias ripCdAsMp3=net_bazzline_media_rip_cd_as_mp3
+fi